changed catalog route layout
